BJS2-9190
- декомпозировал расчет площади треугольника
- поправил по типам
